Add support for `rename` operation on all resoure-based services
Also
- Add way more tests for default resource operations (delete, copy,
move) for all resource-based services.
- Correct some docu flaws

Fixes #204
