+++ src/Client/Generic/ResourceTrait.php
@@ -179,4 +179,20 @@ trait ResourceTrait
             $this->hydratorConfigFactory->createServiceStatusConfig()
         );
     }
+
+    /**
+     * @param int $id
+     * @param string $title
+     *
+     * @return ResourceInformationInterface
+     * @throws EmptyResultException
+     */
+    public function updateTitle(int $id, string $title): ResourceInformationInterface
+    {
+        return $this->responseMapper->getObject(
+            $this->soapClient->rename(['resource_id' => $id, 'name' => $title]),
+            'renameResult',
+            $this->hydratorConfigFactory->createResourceInformationConfig()
+        );
+    }
 }
