+++ src/Client/Mailing/MailingClient.php
@@ -4,6 +4,7 @@ namespace Scn\EvalancheSoapApiConnector\Client\Mailing;
 
 use Scn\EvalancheSoapApiConnector\Client\AbstractClient;
 use Scn\EvalancheSoapApiConnector\Client\ClientInterface;
+use Scn\EvalancheSoapApiConnector\Client\Generic\ResourceTrait;
 use Scn\EvalancheSoapApiConnector\Exception\EmptyResultException;
 use Scn\EvalancheSoapStruct\Struct\Generic\FolderInformationInterface;
 use Scn\EvalancheSoapStruct\Struct\Generic\HashMapInterface;
@@ -30,6 +31,7 @@ use Scn\EvalancheSoapStruct\Struct\Statistic\MailingStatisticInterface;
  */
 final class MailingClient extends AbstractClient implements MailingClientInterface
 {
+    use ResourceTrait;
     const PORTNAME = 'mailing';
     const VERSION = ClientInterface::VERSION_V0;
 
@@ -656,20 +658,6 @@ final class MailingClient extends AbstractClient implements MailingClientInterfa
     }
 
     /**
-     *
-     * @return ResourceTypeInformationInterface[]
-     * @throws EmptyResultException
-     */
-    public function getTypeIds(): array
-    {
-        return $this->responseMapper->getObjects(
-            $this->soapClient->getTypeIds(),
-            'getTypeIdsResult',
-            $this->hydratorConfigFactory->createResourceTypeInformationConfig()
-        );
-    }
-
-    /**
      * @param int $id
      * @param array $attributeTitles
      *
@@ -687,36 +675,6 @@ final class MailingClient extends AbstractClient implements MailingClientInterfa
             ),
             'getUnsubscriptionProfilesResult',
             $this->hydratorConfigFactory->createJobHandleConfig()
-        );
-    }
-
-    /**
-     *
-     * @return ServiceStatusInterface
-     * @throws EmptyResultException
-     */
-    public function getServiceAvailable(): ServiceStatusInterface
-    {
-        return $this->responseMapper->getObject(
-            $this->soapClient->isAlive(),
-            'isAliveResult',
-            $this->hydratorConfigFactory->createServiceStatusConfig()
-        );
-    }
-
-    /**
-     * @param int $id
-     * @param int $folderId
-     *
-     * @return ResourceInformationInterface
-     * @throws EmptyResultException
-     */
-    public function move(int $id, int $folderId): ResourceInformationInterface
-    {
-        return $this->responseMapper->getObject(
-            $this->soapClient->move(['resource_id' => $id, 'category_id' => $folderId]),
-            'moveResult',
-            $this->hydratorConfigFactory->createResourceInformationConfig()
         );
     }
 
@@ -754,22 +712,6 @@ final class MailingClient extends AbstractClient implements MailingClientInterfa
             ),
             'removeArticlesResult',
             $this->hydratorConfigFactory->createMailingArticleConfig()
-        );
-    }
-
-    /**
-     * @param int $id
-     * @param string $title
-     *
-     * @return ResourceInformationInterface
-     * @throws EmptyResultException
-     */
-    public function updateTitle(int $id, string $title): ResourceInformationInterface
-    {
-        return $this->responseMapper->getObject(
-            $this->soapClient->rename(['resource_id' => $id, 'name' => $title]),
-            'renameResult',
-            $this->hydratorConfigFactory->createResourceInformationConfig()
         );
     }
 
@@ -881,111 +823,6 @@ final class MailingClient extends AbstractClient implements MailingClientInterfa
                 ]
             ),
             'setSubjectsResult'
-        );
-    }
-
-    /**
-     * @param int $id
-     *
-     * @return bool
-     * @throws EmptyResultException
-     */
-    public function delete(int $id): bool
-    {
-        return $this->responseMapper->getBoolean(
-            $this->soapClient->delete(['resource_id' => $id]),
-            'deleteResult'
-        );
-    }
-
-    /**
-     * @param int $id
-     * @param int $folderId
-     *
-     * @return ResourceInformationInterface
-     * @throws EmptyResultException
-     */
-    public function copy(int $id, int $folderId): ResourceInformationInterface
-    {
-        return $this->responseMapper->getObject(
-            $this->soapClient->copy(['resource_id' => $id, 'category_id' => $folderId]),
-            'copyResult',
-            $this->hydratorConfigFactory->createResourceInformationConfig()
-        );
-    }
-
-    /**
-     * @param int $id
-     *
-     * @return ResourceInformationInterface[]
-     * @throws EmptyResultException
-     */
-    public function getListByMandatorId(int $id): array
-    {
-        return $this->responseMapper->getObjects(
-            $this->soapClient->getAll(['mandator_id' => $id]),
-            'getAllResult',
-            $this->hydratorConfigFactory->createResourceInformationConfig()
-        );
-    }
-
-    /**
-     * @param int $id
-     *
-     * @return ResourceInformationInterface[]
-     * @throws EmptyResultException
-     */
-    public function getByFolderId(int $id): array
-    {
-        return $this->responseMapper->getObjects(
-            $this->soapClient->getByCategory(['category_id' => $id]),
-            'getByCategoryResult',
-            $this->hydratorConfigFactory->createResourceInformationConfig()
-        );
-    }
-
-    /**
-     * @param int $id
-     *
-     * @return ResourceInformationInterface
-     * @throws EmptyResultException
-     */
-    public function getById(int $id): ResourceInformationInterface
-    {
-        return $this->responseMapper->getObject(
-            $this->soapClient->getById(['resource_id' => $id]),
-            'getByIdResult',
-            $this->hydratorConfigFactory->createResourceInformationConfig()
-        );
-    }
-
-    /**
-     * @param int $id
-     *
-     * @return FolderInformationInterface
-     * @throws EmptyResultException
-     */
-    public function getDefaultFolderByMandatorId(int $id): FolderInformationInterface
-    {
-        return $this->responseMapper->getObject(
-            $this->soapClient->getResourceDefaultCategory(['customer_id' => $id]),
-            'getResourceDefaultCategoryResult',
-            $this->hydratorConfigFactory->createFolderInformationConfig()
-        );
-    }
-
-    /**
-     * @param string $id
-     *
-     * @return ResourceInformationInterface
-     * @throws EmptyResultException
-     */
-    public function getByExternalId(string $id): ResourceInformationInterface
-    {
-        return $this->responseMapper->getObject(
-            $this->soapClient->getByExternalId(['external_id' => $id]),
-            'getByExternalIdResult',
-            $this->hydratorConfigFactory->createResourceInformationConfig()
         );
     }
