+++ src/Client/Mailing/MailingClientInterface.php
@@ -3,6 +3,7 @@
 namespace Scn\EvalancheSoapApiConnector\Client\Mailing;
 
 use Scn\EvalancheSoapApiConnector\Client\ClientInterface;
+use Scn\EvalancheSoapApiConnector\Client\Generic\ResourceTraitInterface;
 use Scn\EvalancheSoapApiConnector\Exception\EmptyResultException;
 use Scn\EvalancheSoapStruct\Struct\Generic\FolderInformationInterface;
 use Scn\EvalancheSoapStruct\Struct\Generic\HashMapInterface;
@@ -27,7 +28,7 @@ use Scn\EvalancheSoapStruct\Struct\Statistic\MailingStatisticInterface;
  *
  * @package Scn\EvalancheSoapApiConnector\Client\Mailing
  */
-interface MailingClientInterface extends ClientInterface
+interface MailingClientInterface extends ResourceTraitInterface, ClientInterface
 {
     /**
      * @param int $id
