+++ src/Client/Mailing/MailingTemplateClient.php
@@ -52,22 +52,6 @@ class MailingTemplateClient extends AbstractClient implements MailingTemplateCli
 
     /**
      * @param int $id
-     * @param string $title
-     *
-     * @return ResourceInformationInterface
-     * @throws EmptyResultException
-     */
-    public function updateTitle(int $id, string $title): ResourceInformationInterface
-    {
-        return $this->responseMapper->getObject(
-            $this->soapClient->rename(['resource_id' => $id, 'name' => $title]),
-            'renameResult',
-            $this->hydratorConfigFactory->createResourceInformationConfig()
-        );
-    }
-
-    /**
-     * @param int $id
      * @param array $articles
      *
      * @return MailingArticleInterface[]
