+++ src/Client/Pool/PoolDataMinerClient.php
@@ -5,8 +5,6 @@ namespace Scn\EvalancheSoapApiConnector\Client\Pool;
 use Scn\EvalancheSoapApiConnector\Client\AbstractClient;
 use Scn\EvalancheSoapApiConnector\Client\ClientInterface;
 use Scn\EvalancheSoapApiConnector\Client\Generic\ResourceTrait;
-use Scn\EvalancheSoapApiConnector\Exception\EmptyResultException;
-use Scn\EvalancheSoapStruct\Struct\Generic\ResourceInformationInterface;
 
 /**
  * Class PoolDataMinerClient
@@ -19,20 +17,4 @@ final class PoolDataMinerClient extends AbstractClient implements PoolDataMinerC
 
     const PORTNAME = 'pooldataminer';
     const VERSION = ClientInterface::VERSION_V0;
-
-    /**
-     * @param int $id
-     * @param string $title
-     *
-     * @return ResourceInformationInterface
-     * @throws EmptyResultException
-     */
-    public function updateTitle(int $id, string $title): ResourceInformationInterface
-    {
-        return $this->responseMapper->getObject(
-            $this->soapClient->rename(['resource_id' => $id, 'name' => $title]),
-            'renameResult',
-            $this->hydratorConfigFactory->createResourceInformationConfig()
-        );
-    }
 }
