+++ tests/Client/Article/ArticleTypeClientTest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Scn\EvalancheSoapApiConnector\Client\Article;
 
 use PHPUnit\Framework\MockObject\MockObject;
+use Scn\EvalancheSoapApiConnector\Client\CommonResourceMethodsTestTrait;
 use Scn\EvalancheSoapApiConnector\EvalancheSoapClient;
 use Scn\EvalancheSoapApiConnector\Extractor\ExtractorInterface;
 use Scn\EvalancheSoapApiConnector\Hydrator\Config\HydratorConfigFactoryInterface;
@@ -24,6 +25,7 @@ use stdClass;
  */
 class ArticleTypeClientTest extends TestCase
 {
+    use CommonResourceMethodsTestTrait;
     /**
      * @var ArticleTypeClient
      */
@@ -65,7 +67,11 @@ class ArticleTypeClientTest extends TestCase
             'removeAttribute',
             'removeAttributeGroup',
             'removeAttributeOption',
-            'updateAttribute'
+            'updateAttribute',
+            'rename',
+            'move',
+            'copy',
+            'delete',
         ]);
         $this->responseMapper = $this->getMockBuilder(ResponseMapperInterface::class)->getMock();
         $this->hydratorConfigFactory = $this->getMockBuilder(HydratorConfigFactoryInterface::class)->getMock();
