+++ tests/Client/Container/ContainerClientTest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Scn\EvalancheSoapApiConnector\Client\Container;
 
 use PHPUnit\Framework\MockObject\MockObject;
+use Scn\EvalancheSoapApiConnector\Client\CommonResourceMethodsTestTrait;
 use Scn\EvalancheSoapApiConnector\EvalancheSoapClient;
 use Scn\EvalancheSoapApiConnector\Extractor\ExtractorInterface;
 use Scn\EvalancheSoapApiConnector\Hydrator\Config\HydratorConfigFactoryInterface;
@@ -23,6 +24,7 @@ use stdClass;
  */
 class ContainerClientTest extends TestCase
 {
+    use CommonResourceMethodsTestTrait;
     /**
      * @var ContainerClient
      */
@@ -56,6 +58,10 @@ class ContainerClientTest extends TestCase
             'update',
             'getDetails',
             'getByContainerTypeId',
+            'rename',
+            'move',
+            'copy',
+            'delete',
         ]);
         $this->responseMapper = $this->getMockBuilder(ResponseMapperInterface::class)->getMock();
         $this->hydratorConfigFactory = $this->getMockBuilder(HydratorConfigFactoryInterface::class)->getMock();
