+++ tests/Client/CouponList/CouponListClientTest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Scn\EvalancheSoapApiConnector\Client\CouponList;
 
 use PHPUnit\Framework\MockObject\MockObject;
+use Scn\EvalancheSoapApiConnector\Client\CommonResourceMethodsTestTrait;
 use Scn\EvalancheSoapApiConnector\EvalancheSoapClient;
 use Scn\EvalancheSoapApiConnector\Extractor\ExtractorInterface;
 use Scn\EvalancheSoapApiConnector\Hydrator\Config\HydratorConfigFactoryInterface;
@@ -17,6 +18,7 @@ use stdClass;
 
 class CouponListClientTest extends TestCase
 {
+    use CommonResourceMethodsTestTrait;
     /**
      * @var CouponListClient
      */
@@ -51,6 +53,10 @@ class CouponListClientTest extends TestCase
             'removeAll',
             'getProfileCouponList',
             'import',
+            'rename',
+            'move',
+            'copy',
+            'delete',
         ]);
         $this->responseMapper = $this->createMock(ResponseMapperInterface::class);
         $this->hydratorConfigFactory = $this->createMock(HydratorConfigFactoryInterface::class);
