+++ tests/Client/Document/DocumentClientTest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Scn\EvalancheSoapApiConnector\Client\Document;
 
 use PHPUnit\Framework\MockObject\MockObject;
+use Scn\EvalancheSoapApiConnector\Client\CommonResourceMethodsTestTrait;
 use Scn\EvalancheSoapApiConnector\EvalancheSoapClient;
 use Scn\EvalancheSoapApiConnector\Extractor\ExtractorInterface;
 use Scn\EvalancheSoapApiConnector\Hydrator\Config\HydratorConfigFactoryInterface;
@@ -24,6 +25,7 @@ use stdClass;
  */
 class DocumentClientTest extends TestCase
 {
+    use CommonResourceMethodsTestTrait;
     /**
      * @var DocumentClient
      */
@@ -63,7 +65,8 @@ class DocumentClientTest extends TestCase
             'getResourceDefaultCategory',
             'getTypeIds',
             'move',
-            'isAlive'
+            'isAlive',
+            'rename',
         ]);
         $this->responseMapper = $this->getMockBuilder(ResponseMapperInterface::class)->getMock();
         $this->hydratorConfigFactory = $this->getMockBuilder(HydratorConfigFactoryInterface::class)->getMock();
@@ -103,50 +106,6 @@ class DocumentClientTest extends TestCase
             ResourceInformationInterface::class,
             $this->subject->create($title, $folderId)
         );
-    }
-
-    public function testCopyCanReturnInstanceOfResourceInformation()
-    {
-        $id = 456;
-        $folderId = 123;
-
-        $config = $this->getMockBuilder(HydratorConfigInterface::class)->getMock();
-        $object = $this->getMockBuilder(ResourceInformationInterface::class)->getMock();
-
-        $response = new stdClass();
-        $response->copyResult = $object;
-
-        $this->hydratorConfigFactory->expects($this->once())->method('createResourceInformationConfig')->willReturn($config);
-        $this->soapClient->expects($this->once())->method('copy')->with([
-            'resource_id' => $id,
-            'category_id' => $folderId
-        ])->willReturn($response);
-        $this->responseMapper->expects($this->once())->method('getObject')->with(
-            $response,
-            'copyResult',
-            $config
-        )->willReturn($response->copyResult);
-
-        $this->assertInstanceOf(
-            ResourceInformationInterface::class,
-            $this->subject->copy($id, $folderId)
-        );
-    }
-
-    public function testDeleteCanReturnBoolean()
-    {
-        $id = 56;
-
-        $response = new stdClass();
-        $response->deleteResult = true;
-
-        $this->soapClient->expects($this->once())->method('delete')->with(['resource_id' => $id])->willReturn($response);
-        $this->responseMapper->expects($this->once())->method('getBoolean')->with(
-            $response,
-            'deleteResult'
-        )->willReturn($response->deleteResult);
-
-        $this->assertTrue($this->subject->delete($id));
     }
 
     public function testGetListByMandatorIdCanReturnArrayOfResourceInformation()
