+++ tests/Client/Pool/PoolDataMinerClientTest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Scn\EvalancheSoapApiConnector\Client\Pool;
 
 use PHPUnit\Framework\MockObject\MockObject;
+use Scn\EvalancheSoapApiConnector\Client\CommonResourceMethodsTestTrait;
 use Scn\EvalancheSoapApiConnector\EvalancheSoapClient;
 use Scn\EvalancheSoapApiConnector\Extractor\ExtractorInterface;
 use Scn\EvalancheSoapApiConnector\Hydrator\Config\HydratorConfigFactoryInterface;
@@ -21,6 +22,7 @@ use stdClass;
  */
 class PoolDataMinerClientTest extends TestCase
 {
+    use CommonResourceMethodsTestTrait;
     /**
      * @var PoolDataMinerClient
      */
@@ -48,7 +50,12 @@ class PoolDataMinerClientTest extends TestCase
 
     public function setUp(): void
     {
-        $this->soapClient = $this->getWsdlMock(['rename']);
+        $this->soapClient = $this->getWsdlMock([
+            'rename',
+            'copy',
+            'delete',
+            'move',
+        ]);
         $this->responseMapper = $this->getMockBuilder(ResponseMapperInterface::class)->getMock();
         $this->hydratorConfigFactory = $this->getMockBuilder(HydratorConfigFactoryInterface::class)->getMock();
         $this->extractor = $this->getMockBuilder(ExtractorInterface::class)->getMock();
