+++ tests/Client/Smartlink/SmartLinkClientTest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Scn\EvalancheSoapApiConnector\Client\Smartlink;
 
 use PHPUnit\Framework\MockObject\MockObject;
+use Scn\EvalancheSoapApiConnector\Client\CommonResourceMethodsTestTrait;
 use Scn\EvalancheSoapApiConnector\EvalancheSoapClient;
 use Scn\EvalancheSoapApiConnector\Extractor\ExtractorInterface;
 use Scn\EvalancheSoapApiConnector\Hydrator\Config\HydratorConfigFactoryInterface;
@@ -22,6 +23,7 @@ use stdClass;
  */
 class SmartLinkClientTest extends TestCase
 {
+    use CommonResourceMethodsTestTrait;
     /**
      * @var SmartLinkClient
      */
@@ -54,7 +56,11 @@ class SmartLinkClientTest extends TestCase
             'getTrackingUrls',
             'getLinkConfigurations',
             'setLinkConfigurations',
-            'createLinkConfigurations'
+            'createLinkConfigurations',
+            'copy',
+            'rename',
+            'move',
+            'delete',
         ]);
         $this->responseMapper = $this->getMockBuilder(ResponseMapperInterface::class)->getMock();
         $this->hydratorConfigFactory = $this->getMockBuilder(HydratorConfigFactoryInterface::class)->getMock();
